scripts/rag/requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pandas==2.1.4
numpy==1.24.3
tiktoken==0.5.2
pydantic==2.5.2
uvicorn==0.24.0
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2